+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.3] - breakout changes
+
+- use actions widgets list to add buttons.
+- add out of the box customized buttons widgets.
+
+
 ## [0.0.2] - fix single button padding bug and edit ReadMe.
 
 ## [0.0.1] - first release.